Created locked_landing_pages service.
