+++ web/modules/custom/server_general/server_general.module
@@ -110,9 +110,12 @@ function server_general_node_access(NodeInterface $entity, string $op, AccountIn
     return AccessResult::neutral();
   }
 
-  $main_settings = server_general_get_main_settings();
+  /** @var \Drupal\server_general\LockedLandingPages $locked_lp_service */
+  $locked_lp_service = \Drupal::service('server_general.locked_landing_pages');
 
-  if ($op === 'delete' && $main_settings instanceof ConfigPages && server_general_is_node_locked($entity)) {
+  $main_settings = $locked_lp_service->getMainSettings();
+
+  if ($op === 'delete' && $main_settings instanceof ConfigPages && $locked_lp_service->isNodeLocked($entity)) {
     // Add cache tags to the cache metadata.
     $cache_tags = $main_settings->getCacheTags();
     return AccessResult::forbidden()->addCacheableDependency($entity)->addCacheTags($cache_tags);
@@ -125,7 +128,10 @@ function server_general_node_access(NodeInterface $entity, string $op, AccountIn
  * Implements hook_entity_predelete().
  */
 function server_general_node_predelete(NodeInterface $entity) {
-  if (!server_general_is_node_locked($entity)) {
+  /** @var \Drupal\server_general\LockedLandingPages $locked_lp_service */
+  $locked_lp_service = \Drupal::service('server_general.locked_landing_pages');
+
+  if (!$locked_lp_service->isNodeLocked($entity)) {
     return;
   }
 
@@ -136,7 +142,10 @@ function server_general_node_predelete(NodeInterface $entity) {
  * Implements hook_node_presave().
  */
 function server_general_node_presave(NodeInterface $entity) {
-  if (!server_general_is_node_locked($entity)) {
+  /** @var \Drupal\server_general\LockedLandingPages $locked_lp_service */
+  $locked_lp_service = \Drupal::service('server_general.locked_landing_pages');
+
+  if (!$locked_lp_service->isNodeLocked($entity)) {
     return;
   }
 
@@ -170,9 +179,11 @@ function server_general_form_alter(array &$form, FormStateInterface $form_state)
   if ($form['#form_id'] !== "node_{$entity->bundle()}_edit_form") {
     return;
   }
+  /** @var \Drupal\server_general\LockedLandingPages $locked_lp_service */
+  $locked_lp_service = \Drupal::service('server_general.locked_landing_pages');
 
   // Check if node is locked or not.
-  if (!server_general_is_node_locked($entity)) {
+  if (!$locked_lp_service->isNodeLocked($entity)) {
     return;
   }
 
@@ -188,10 +199,12 @@ function server_general_entity_operation_alter(array &$operations, EntityInterfa
   if ($entity->getEntityType()->id() !== 'node') {
     return;
   }
+  /** @var \Drupal\server_general\LockedLandingPages $locked_lp_service */
+  $locked_lp_service = \Drupal::service('server_general.locked_landing_pages');
 
   // Check if node is locked or not.
   /** @var \Drupal\node\NodeInterface $entity */
-  if (!server_general_is_node_locked($entity)) {
+  if (!$locked_lp_service->isNodeLocked($entity)) {
     return;
   }
 
@@ -220,43 +233,17 @@ function server_general_menu_local_tasks_alter(array &$data, string $route_name)
   if (!$entity instanceof NodeInterface) {
     return;
   }
+
+  /** @var \Drupal\server_general\LockedLandingPages $locked_lp_service */
+  $locked_lp_service = \Drupal::service('server_general.locked_landing_pages');
+
   // Check if node is locked or not.
   /** @var \Drupal\node\NodeInterface $entity */
-  if (!server_general_is_node_locked($entity)) {
+  if (!$locked_lp_service->isNodeLocked($entity)) {
     return;
   }
 
   unset($data['tabs'][0]['entity.node.delete_form']);
-}
-
-/**
- * Returns list of restricted node ids from config pages.
- *
- * @return array
- *   List of ids.
- */
-function server_general_get_list_of_restricted_nodes() {
-  $main_settings = server_general_get_main_settings();
-
-  if (!$main_settings instanceof ConfigPages) {
-    return [];
-  }
-
-  $locked_nodes = $main_settings->get('field_locked_landing_pages')->getValue();
-
-  return array_column($locked_nodes, 'target_id');
-}
-
-/**
- * Checks if current node is locked.
- *
- * @return bool
- *   Returns TRUE if node is locked.
- */
-function server_general_is_node_locked(NodeInterface $node) {
-  $restricted_nodes = server_general_get_list_of_restricted_nodes();
-
-  return in_array($node->id(), $restricted_nodes);
 }
 
 /**
@@ -268,18 +255,4 @@ function server_general_preprocess_media_oembed_iframe(array &$variables) {
   if (strpos((string) $variables['media'], 'youtube.com') !== FALSE) {
     $variables['media'] = IFrameMarkup::create(str_replace('youtube.com/', 'youtube-nocookie.com/', $variables['media']));
   }
-}
-
-/**
- * Returns the 'main_settings' config page.
- *
- * @return \Drupal\Core\Entity\ContentEntityInterface|null
- *   The 'main_settings' config page entity or null if not found.
- */
-function server_general_get_main_settings() {
-  /** @var \Drupal\config_pages\ConfigPagesStorage $config_pages_storage */
-  $config_pages_storage = \Drupal::entityTypeManager()->getStorage('config_pages');
-  /** @var \Drupal\Core\Entity\ContentEntityInterface|null $main_settings */
-  $main_settings = $config_pages_storage->load('main_settings');
-  return $main_settings;
 }
